Worked on the employee and admin entities
I've worked on the employee and admin entities.
Employees can now be given orderlists which they can then prepare and process!
I also added a paid (boolean) status to orders! The webshop can check on their own bankaccount if a transfer has been made and then change this status to paid.
I also fixed some bugs! One of them was in the UserServiceImpl, I added some if statements to prevent a nullpointerexception.